refactor(cli): move chip modules into uniqc/cli/ and add --version flag
- Move chip_info.py, chip_cache.py, chip_service.py from uniqc/ to uniqc/cli/
  to group CLI-adjacent modules together
- Add --version/-V flag to the main uniqc CLI using typer callback
- Update all import paths across backend.py and platform adapters
