+++ uniqc/task/adapters/ibm_adapter.py
@@ -318,7 +318,7 @@ class IBMAdapter(QuantumAdapter):
         from datetime import datetime, timezone
 
         from uniqc.backend_info import Platform, QubitTopology
-        from uniqc.chip_info import (
+        from uniqc.cli.chip_info import (
             ChipCharacterization,
             ChipGlobalInfo,
             SingleQubitData,
