+++ uniqc/task/adapters/originq_adapter.py
@@ -416,7 +416,7 @@ class OriginQAdapter(QuantumAdapter):
         from datetime import datetime, timezone
 
         from uniqc.backend_info import Platform, QubitTopology
-        from uniqc.chip_info import (
+        from uniqc.cli.chip_info import (
             ChipCharacterization,
             ChipGlobalInfo,
             SingleQubitData,
